python/requirements.txt
MetaTrader5==5.0.45
requests==2.32.3
python-dotenv==1.0.1
